huggingface/requirements.txt
gradio
ultralytics
Pillow
firebase-admin
numpy